1.3.2

Former-commit-id: 38101c9691b2606e526d68e2cfcae7c467517bd1

+++ src/ru/meefik/linuxdeploy/MainActivity.java
@@ -322,8 +322,8 @@ public class MainActivity extends Activity implements OnClickListener {
 				ssh = "  SSH: " + PrefStore.SSH_PORT;
 			if (PrefStore.VNC_START == "y")
 				vnc = "  VNC: "
-						+ String.valueOf(5900 + Integer
-								.parseInt(PrefStore.VNC_DISPLAY));
+						+ String.valueOf(5900 + (int) Double
+								.parseDouble(PrefStore.VNC_DISPLAY));
 			titleMsg += "  [ " + myIP + ssh + vnc + " ]";
 		}
